refactor: extract integrity check and usage stats into dev/check.rs
Move the cluster-usage accounting (qcow2_cluster_usage and its
add_*_clusters helpers) and the integrity checker (check and its
check_*/cluster_is_allocated helpers) into a new dev/check.rs submodule
as an impl extension of Qcow2Dev. No visibility changes needed. No logic
changes.
